third_party/WebKit/LayoutTests/inspector/console/console-log-linkify-links-expected.txt

CONSOLE MESSAGE: line 6: http://www.chromium.org/
CONSOLE MESSAGE: line 7: follow http://www.chromium.org/
CONSOLE MESSAGE: line 8: string
CONSOLE MESSAGE: line 9: 123
Test that console.log() would linkify the links. Bug 231074.

http://www.chromium.org/ console-log-linkify-links.html:6 console-message-text source-code > webkit-html-external-link > console-message-url webkit-html-resource-link
follow http://www.chromium.org/ console-log-linkify-links.html:7 console-message-text source-code > webkit-html-external-link > console-message-url webkit-html-resource-link
string http://www.chromium.org/ console-log-linkify-links.html:8 console-message-text source-code > webkit-html-external-link > console-message-url webkit-html-resource-link
123 "http://www.chromium.org/" console-log-linkify-links.html:9 console-message-text source-code > console-formatted-number source-code > source-code > console-formatted-string source-code > webkit-html-external-link > console-message-url webkit-html-resource-link
